requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
contextlib2==21.6.0
Django==5.2.7
django-cors-headers==4.9.0
django-ninja==1.1.0
django-ninja-extra==0.20.7
django-ninja-jwt==5.3.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fire==0.7.1
idna==3.11
injector==0.22.0
markdown-it-py==4.0.0
mdurl==0.1.2
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.3
rav==0.1.1
rich==14.2.0
sqlparse==0.5.3
termcolor==3.1.0
typing-inspection==0.4.2
